requirements.txt
# python==3.10
# CUDA==11.8
# NVIDIA GeForce RTX 4060

torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
scikit-learn==1.5.1
numpy==1.26.4
pillow==11.0.0
scipy==1.14.1
tabulate==0.9.0